requirements.txt

langchain>=1.2.13
langchain-core>=1.2.20
langchain-aws>=1.4.1
langchain-community>=0.4.1
langchain-chroma>=1.1.0
langchain-text-splitters>=1.1.1
langchain-mcp-adapters>=0.2.2
langgraph>=1.1.3
langsmith>=0.7.22
chromadb>=1.5.5
mcp>=1.26.0
fastapi>=0.135.1
uvicorn>=0.42.0
python-dotenv>=1.2.2
boto3>=1.42.73
pydantic>=2.12.5
pypdf>=6.9.1
streamlit>=1.55.0